gtm-ai-temp/requirements.txt
aiohttp
beautifulsoup4
python-dotenv
openai>=1.0.0
certifi
